namada_token::masp::shielded_wallet
Trait ShieldedQueries

pub trait ShieldedQueries<U>: Deref<Target = ShieldedWallet<U>> + DerefMut
where
    U: ShieldedUtils + MaybeSend + MaybeSync,
{
    // Required methods
    async fn query_native_token<C>(client: &C) -> Result<Address, Report>
       where C: Client + Sync;
    async fn query_denom<C>(client: &C, token: &Address) -> Option<Denomination>
       where C: Client + Sync;
    async fn query_conversion<C>(
        client: &C,
        asset_type: AssetType,
    ) -> Option<(Address, Denomination, MaspDigitPos, MaspEpoch, ValueSum<AssetType, i128>, MerklePath<Node>)>
       where C: Client + Sync;
    async fn query_block<C>(client: &C) -> Result<Option<u64>, Report>
       where C: Client + Sync;
    async fn query_max_block_time_estimate<C>(
        client: &C,
    ) -> Result<DurationSecs, Report>
       where C: Client + Sync;
    async fn query_masp_epoch<C>(client: &C) -> Result<MaspEpoch, Report>
       where C: Client + Sync;
}

A trait that allows downstream types specify how a shielded wallet should interact / query a node.

Required Methods

async fn query_native_token<C>(client: &C) -> Result<Address, Report>
where C: Client + Sync,

Get the address of the native token

async fn query_denom<C>(client: &C, token: &Address) -> Option<Denomination>
where C: Client + Sync,

Query the denomination of a token type

async fn query_conversion<C>( client: &C, asset_type: AssetType, ) -> Option<(Address, Denomination, MaspDigitPos, MaspEpoch, ValueSum<AssetType, i128>, MerklePath<Node>)>
where C: Client + Sync,

Query for converting assets across epochs

async fn query_block<C>(client: &C) -> Result<Option<u64>, Report>
where C: Client + Sync,

Get the last block height

async fn query_max_block_time_estimate<C>( client: &C, ) -> Result<DurationSecs, Report>
where C: Client + Sync,

Get the upper limit on the time to make a new block

async fn query_masp_epoch<C>(client: &C) -> Result<MaspEpoch, Report>
where C: Client + Sync,

Query the MASP epoch

Object Safety

This trait is not object safe.