namada_token::masp
Trait ShieldedUtils

pub trait ShieldedUtils:
    Sized
    + BorshDeserialize
    + BorshSerialize
    + Default
    + Clone {
    // Required methods
    fn local_tx_prover(&self) -> LocalTxProver;
    fn load<'life0, 'life1, 'async_trait, U>(
        &'life0 self,
        ctx: &'life1 mut ShieldedWallet<U>,
        force_confirmed: bool,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             U: 'async_trait + ShieldedUtils + MaybeSend,
             Self: 'async_trait;
    fn save<'life0, 'life1, 'async_trait, U>(
        &'life0 self,
        ctx: &'life1 ShieldedWallet<U>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             U: 'async_trait + ShieldedUtils + MaybeSync,
             Self: 'async_trait;
    fn cache_save<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _cache: &'life1 DispatcherCache,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn cache_load<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<DispatcherCache, Error>> + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
}

Abstracts platform specific details away from the logic of shielded pool operations.

Required Methods

fn local_tx_prover(&self) -> LocalTxProver

Get a MASP transaction prover

fn load<'life0, 'life1, 'async_trait, U>( &'life0 self, ctx: &'life1 mut ShieldedWallet<U>, force_confirmed: bool, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, U: 'async_trait + ShieldedUtils + MaybeSend, Self: 'async_trait,

Load up the currently saved ShieldedContext

fn save<'life0, 'life1, 'async_trait, U>( &'life0 self, ctx: &'life1 ShieldedWallet<U>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, U: 'async_trait + ShieldedUtils + MaybeSync, Self: 'async_trait,

Save the given ShieldedContext for future loads

fn cache_save<'life0, 'life1, 'async_trait>( &'life0 self, _cache: &'life1 DispatcherCache, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Save a cache of data as part of shielded sync if that process gets interrupted.

fn cache_load<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<DispatcherCache, Error>> + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Load a cache of data as part of shielded sync if that process gets interrupted.

Object Safety

This trait is not object safe.