namada_token::masp
Type Alias SpentNotesTracker

pub type SpentNotesTracker = IndexMap<ViewingKey, IndexSet<usize>>;

A type tracking the notes used to construct a shielded transfer. Used to avoid reusing the same notes multiple times which would lead to an invalid transaction

Aliased Type

struct SpentNotesTracker { /* private fields */ }