namada_token::masp::utils
Trait MaspClient

pub trait MaspClient: Clone {
    type Error: Error + Send + Sync + 'static;

    // Required methods
    async fn last_block_height(
        &self,
    ) -> Result<Option<BlockHeight>, Self::Error>;
    async fn fetch_shielded_transfers(
        &self,
        from: BlockHeight,
        to: BlockHeight,
    ) -> Result<Vec<(IndexedTx, Transaction)>, Self::Error>;
    fn capabilities(&self) -> MaspClientCapabilities;
    async fn fetch_commitment_tree(
        &self,
        height: BlockHeight,
    ) -> Result<CommitmentTree<Node>, Self::Error>;
    async fn fetch_note_index(
        &self,
        height: BlockHeight,
    ) -> Result<BTreeMap<IndexedTx, usize>, Self::Error>;
    async fn fetch_witness_map(
        &self,
        height: BlockHeight,
    ) -> Result<IndexMap<usize, IncrementalWitness<Node>>, Self::Error>;
}

This abstracts away the implementation details of how shielded-sync fetches the necessary data from a remote server.

Required Associated Types

type Error: Error + Send + Sync + 'static

Error type returned by the methods of this trait

Required Methods

async fn last_block_height(&self) -> Result<Option<BlockHeight>, Self::Error>

Return the last block height we can retrieve data from.

async fn fetch_shielded_transfers( &self, from: BlockHeight, to: BlockHeight, ) -> Result<Vec<(IndexedTx, Transaction)>, Self::Error>

Fetch shielded transfers from blocks heights in the range [from, to], keeping track of progress through progress. The fetched transfers are sent over to a separate worker through tx_sender.

fn capabilities(&self) -> MaspClientCapabilities

Return the capabilities of this client.

async fn fetch_commitment_tree( &self, height: BlockHeight, ) -> Result<CommitmentTree<Node>, Self::Error>

Fetch the commitment tree of height height.

async fn fetch_note_index( &self, height: BlockHeight, ) -> Result<BTreeMap<IndexedTx, usize>, Self::Error>

Fetch the tx notes map of height height.

async fn fetch_witness_map( &self, height: BlockHeight, ) -> Result<IndexMap<usize, IncrementalWitness<Node>>, Self::Error>

Fetch the witness map of height height.

Object Safety

This trait is not object safe.