namada_token
Module storage_key

Token storage keys

Constants

BALANCE_STORAGE_KEY

Key segment for a balance key

DENOM_STORAGE_KEY

Key segment for a denomination key

MINTED_STORAGE_KEY

Key segment for minted balance

MINTER_STORAGE_KEY

Key segment for multitoken minter

PARAMETERS_STORAGE_KEY

Key segment for token parameters

Functions

balance_key

Obtain a storage key for user’s balance.

balance_prefix

Obtain a storage key prefix for all users’ balances.

denom_key

Obtain a storage key denomination of a token.

is_any_minted_balance_key

Check if the given storage key is for total supply of a unspecified token. If it is, returns the token.

is_any_minter_key

Check if the given storage key is for a minter of a unspecified token. If it is, returns the token.

is_any_token_balance_key

Check if the given storage key is a balance key for an unspecified token. If it is, return the token and owner address.

is_any_token_parameter_key

Check if the given storage key is a parameter key for an unspecified token. If it is, return the token address.

is_balance_key

Check if the given storage key is a balance key for the given token. If it is, return the owner. For minted balances, use is_any_minted_balance_key().

is_denom_key

Check if the given storage key is a denomination key for the given token.

is_masp_commitment_anchor_key

Check if the given key is a masp commitment anchor

is_masp_key

Check if the given storage key is a masp key

is_masp_nullifier_key

Check if the given storage key is a masp nullifier key

is_masp_token_map_key

Check if the given storage key is a masp token map key

is_masp_transfer_key

Check if the given storage key is allowed to be touched by a masp transfer

is_multitoken_key

Check if a key is part of the multitoken vp sub storage

key_of_token

Gets the key for the given token address, error with the given message to expect if the key is not in the address

masp_assets_hash_key

Get the key for the masp assets’ hash

masp_commitment_anchor_key

Get a key for a masp commitment tree anchor

masp_commitment_tree_key

Get the key for the masp commitment tree

masp_convert_anchor_key

Get the key for the masp convert tree anchor

masp_kd_gain_key

Obtain the nominal derivative key for the given token

masp_kp_gain_key

Obtain the nominal proportional key for the given token

masp_last_inflation_key

Obtain the storage key for the last inflation of a token

masp_last_locked_amount_key

Obtain the storage key for the last locked amount of a token

masp_locked_amount_target_key

Obtain the locked target amount key for the given token

masp_max_reward_rate_key

The max reward rate key for the given token

masp_nullifier_key

Get a key for a masp nullifier

masp_token_map_key

Get the key for the masp token map

masp_total_rewards

The max reward rate key for the given token

minted_balance_key

Obtain a storage key for the minted multitoken balance.

minter_key

Obtain a storage key for the multitoken minter.

parameter_prefix

Obtain a storage key prefix for token parameters.