namada_token
Trait Write

pub trait Write<S>: Read<S> {
    // Required methods
    fn transfer(
        storage: &mut S,
        token: &Address,
        src: &Address,
        dest: &Address,
        amount: Amount,
    ) -> Result<(), Error>;
    fn burn_tokens(
        storage: &mut S,
        token: &Address,
        source: &Address,
        amount: Amount,
    ) -> Result<(), Error>;
    fn credit_tokens(
        storage: &mut S,
        token: &Address,
        dest: &Address,
        amount: Amount,
    ) -> Result<(), Error>;
}

Abstract token storage write interface

Required Methods

fn transfer( storage: &mut S, token: &Address, src: &Address, dest: &Address, amount: Amount, ) -> Result<(), Error>

Transfer token from src to dest. Returns an Err if src has insufficient balance or if the transfer the dest would overflow (This can only happen if the total supply doesn’t fit in token::Amount).

fn burn_tokens( storage: &mut S, token: &Address, source: &Address, amount: Amount, ) -> Result<(), Error>

Burn a specified amount of tokens from some address. If the burn amount is larger than the total balance of the given address, then the remaining balance is burned. The total supply of the token is properly adjusted.

fn credit_tokens( storage: &mut S, token: &Address, dest: &Address, amount: Amount, ) -> Result<(), Error>

Credit tokens to an account, to be used only by protocol. In transactions, this would get rejected by the default vp_token.

Object Safety

This trait is not object safe.