namada_token
Module validation

MASP verification wrappers.

Structs

PVKs

MASP verifying keys

PartialAuthorized

Represents an authorization where the Sapling bundle is authorized and the transparent bundle is unauthorized.

Constants

CONVERT_NAME

Convert circuit name

ENV_VAR_MASP_PARAMS_DIR

Env var to point to a dir with MASP parameters. When not specified, the default OS specific path is used.

OUTPUT_NAME

Output circuit name

SPEND_NAME

Spend circuit name

Functions

get_params_dir

Get the path to MASP parameters from ENV_VAR_MASP_PARAMS_DIR env var or use the default.

partial_deauthorize

Partially deauthorize the transparent bundle

preload_verifying_keys

Make sure the MASP params are present and load verifying keys into memory

verify_shielded_tx

Verify a shielded transaction.